Address an issue that led to beta for cluster size to be misestimated
In the original data, zero time use was recorded as missing, which is fine. However, in some cases we need to sum to time variables, one of which could be missing, which would lead to a missing sum. To address this, I set missing values to zero, but and in taking logs was careful not to compute log(0). However, this leads to the inclusion of misleading values in analysis that, in some cases, results in quite severe misestimation. In this version, we set back to missing times for activitvies that were not observed.

Observations of zero time use were being set to values with large magnitude (on the log scale) in an attempt to avoid taking logs of zero. It is important